Yay, got a first cut at something that finds the dictionary of GAAP data that actually corresponds to the current quarter reported in the form 10.
Discovered that I needed to be using datetime instead of time in order to
compute differences between the start and the end of the period. I'm still
concerned that something's going wrong there, because I'm seeing lots of
272 day-periods, which I suspect are wrong.

Started to rework the context code to be more robust to errors. Still need
to pull it out into a separate method.

Also still need to convert all the methods in this file into a class that
other classes can call.